Revert conflicting changes
