+++ src/main/java/org/folio/innreach/domain/service/impl/AgencyServiceImpl.java
@@ -1,7 +1,6 @@
 package org.folio.innreach.domain.service.impl;
 
 import static java.util.Collections.emptyList;
-import static java.util.Optional.ofNullable;
 import static org.apache.commons.collections4.CollectionUtils.isNotEmpty;
 
 import static org.folio.innreach.external.dto.InnReachResponse.OK_STATUS;
@@ -9,11 +8,8 @@ import static org.folio.innreach.util.ListUtils.flatMapItems;
 import static org.folio.innreach.util.ListUtils.mapItemsWithFilter;
 import static org.folio.innreach.util.ListUtils.toStream;
 
-import java.util.Collections;
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
-import java.util.UUID;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
@@ -29,7 +25,6 @@ import org.folio.innreach.dto.Agency;
 import org.folio.innreach.dto.CentralServerAgenciesDTO;
 import org.folio.innreach.dto.CentralServerDTO;
 import org.folio.innreach.dto.InnReachResponseDTO;
-import org.folio.innreach.dto.LocalServer;
 import org.folio.innreach.dto.LocalServerAgenciesDTO;
 import org.folio.innreach.external.exception.InnReachException;
 import org.folio.innreach.external.service.InnReachExternalService;
@@ -55,28 +50,31 @@ public class AgencyServiceImpl implements AgencyService {
     var agencies = mapItemsWithFilter(servers, this::retrieveAllAgenciesFromCentralServer, Objects::nonNull);
 
     return new CentralServerAgenciesDTO()
-      .centralServerAgencies(agencies)
-      .totalRecords(agencies.size());
-  }
-
-  @Override
-  public List<LocalServer> getLocalServers(UUID centralServerId) {
-    var server = centralServerService.getCentralServer(centralServerId);
-    return ofNullable(retrieveLocalServerAgencies(server))
-      .map(LocalServerAgenciesDTO::getLocalServerList)
-      .orElse(Collections.emptyList());
+        .centralServerAgencies(agencies)
+        .totalRecords(agencies.size());
   }
 
   private AgenciesPerCentralServerDTO retrieveAllAgenciesFromCentralServer(CentralServerDTO centralServer) {
-    var lsaResponse = retrieveLocalServerAgencies(centralServer);
-    if (lsaResponse == null) return null;
+    log.info("Retrieving local server agencies from central server: code = {}", centralServer.getCentralServerCode());
+
+    LocalServerAgenciesDTO lsaResponse;
+    try {
+      String response = innReachService.callInnReachApi(centralServer.getId(), INN_REACH_LOCAL_SERVERS_URI);
+
+      lsaResponse = jsonHelper.fromJson(response, LocalServerAgenciesDTO.class);
+    } catch (InnReachException | BadCredentialsException | IllegalStateException e) {
+      log.warn("Failed to get local server agencies from central server: code = {}",
+          centralServer.getCentralServerCode(), e);
+
+      return null;
+    }
 
     List<Agency> agencies = emptyList();
     if (isOk(lsaResponse)) {
       agencies = flatMapItems(lsaResponse.getLocalServerList(), localServer -> toStream(localServer.getAgencyList()));
     } else {
       log.warn("Failed to get local server agencies from central server: code = {}. Inn-reach response: {}",
-        centralServer.getCentralServerCode(), lsaResponse);
+          centralServer.getCentralServerCode(), lsaResponse);
     }
 
     log.info("Number of agencies received: {}", agencies.size());
@@ -84,25 +82,11 @@ public class AgencyServiceImpl implements AgencyService {
     return isNotEmpty(agencies) ? createAgencies(centralServer, agencies) : null;
   }
 
-  private LocalServerAgenciesDTO retrieveLocalServerAgencies(CentralServerDTO centralServer) {
-    log.info("Retrieving local server agencies from central server: code = {}", centralServer.getCentralServerCode());
-    try {
-      String response = innReachService.callInnReachApi(centralServer.getId(), INN_REACH_LOCAL_SERVERS_URI);
-
-      return jsonHelper.fromJson(response, LocalServerAgenciesDTO.class);
-    } catch (InnReachException | BadCredentialsException | IllegalStateException e) {
-      log.warn("Failed to get local server agencies from central server: code = {}",
-        centralServer.getCentralServerCode(), e);
-
-      return null;
-    }
-  }
-
   private AgenciesPerCentralServerDTO createAgencies(CentralServerDTO centralServer, List<Agency> agencies) {
     return new AgenciesPerCentralServerDTO()
-      .centralServerId(centralServer.getId())
-      .centralServerCode(centralServer.getCentralServerCode())
-      .agencies(agencies);
+        .centralServerId(centralServer.getId())
+        .centralServerCode(centralServer.getCentralServerCode())
+        .agencies(agencies);
   }
 
   private static boolean isOk(InnReachResponseDTO innReachResponse) {
