+++ src/test/java/org/folio/innreach/controller/CentralServerAgencyControllerTest.java
@@ -1,6 +1,7 @@
 package org.folio.innreach.controller;
 
 import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
+import static com.github.tomakehurst.wiremock.client.WireMock.equalTo;
 import static com.github.tomakehurst.wiremock.client.WireMock.stubFor;
 import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
 import static org.springframework.test.context.jdbc.Sql.ExecutionPhase.AFTER_TEST_METHOD;
@@ -12,12 +13,16 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.folio.innreach.external.InnReachHeaders.X_TO_CODE;
 import static org.folio.innreach.fixture.TestUtil.readFile;
 
+import java.util.Collections;
 import java.util.Map;
 
+import com.github.tomakehurst.wiremock.client.MappingBuilder;
 import com.github.tomakehurst.wiremock.client.WireMock;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.MediaType;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.SqlMergeMode;
 
@@ -27,17 +32,16 @@ import org.folio.innreach.fixture.CentralServerFixture;
 import org.folio.innreach.repository.CentralServerRepository;
 
 @Sql(
-  scripts = {
-    "classpath:db/central-server/clear-central-server-tables.sql"},
-  executionPhase = AFTER_TEST_METHOD
+    scripts = {
+        "classpath:db/central-server/clear-central-server-tables.sql"},
+    executionPhase = AFTER_TEST_METHOD
 )
 @SqlMergeMode(MERGE)
-public
 class CentralServerAgencyControllerTest extends BaseApiControllerTest {
 
   private static final String CS_TEST_CODE1 = "test1";
   private static final String CS_TEST_CODE2 = "test2";
-  public static final String INNREACH_LOCALSERVERS_URL = "/innreach/v2/contribution/localservers";
+  private static final String INNREACH_LOCALSERVERS_URL = "/innreach/v2/contribution/localservers";
   private static final String AGENCIES_URL = "/inn-reach/central-servers/agencies";
 
   @Autowired
@@ -51,9 +55,9 @@ class CentralServerAgencyControllerTest extends BaseApiControllerTest {
     stubGet(INNREACH_LOCALSERVERS_URL, "agency-codes/d2ir-local-servers-response-01.json");
 
     mockMvc.perform(get(AGENCIES_URL))
-      .andExpect(status().isOk())
-      .andExpect(content()
-        .json(readTemplate("cs-agencies-single-server-response.json", cs.getId(), CS_TEST_CODE1)));
+        .andExpect(status().isOk())
+        .andExpect(content()
+            .json(readTemplate("cs-agencies-single-server-response.json", cs.getId(), CS_TEST_CODE1)));
   }
 
   @Test
@@ -62,22 +66,22 @@ class CentralServerAgencyControllerTest extends BaseApiControllerTest {
     var cs2 = createCentralServer(CS_TEST_CODE2);
 
     stubGet(INNREACH_LOCALSERVERS_URL, xToCodeHeader(CS_TEST_CODE1),
-      "agency-codes/d2ir-local-servers-response-01.json");
+        "agency-codes/d2ir-local-servers-response-01.json");
     stubGet(INNREACH_LOCALSERVERS_URL, xToCodeHeader(CS_TEST_CODE2),
-      "agency-codes/d2ir-local-servers-response-02.json");
+        "agency-codes/d2ir-local-servers-response-02.json");
 
     mockMvc.perform(get(AGENCIES_URL))
-      .andExpect(status().isOk())
-      .andExpect(content()
-        .json(readTemplate("cs-agencies-two-server-response.json", cs1.getId(), CS_TEST_CODE1,
-          cs2.getId(), CS_TEST_CODE2)));
+        .andExpect(status().isOk())
+        .andExpect(content()
+            .json(readTemplate("cs-agencies-two-server-response.json", cs1.getId(), CS_TEST_CODE1,
+                cs2.getId(), CS_TEST_CODE2)));
   }
 
   @Test
   void returnEmptyAgencyCodeListIfNoCentralServers() throws Exception {
     mockMvc.perform(get(AGENCIES_URL))
-      .andExpect(status().isOk())
-      .andExpect(content().json(readTemplate("cs-agencies-empty-response.json")));
+        .andExpect(status().isOk())
+        .andExpect(content().json(readTemplate("cs-agencies-empty-response.json")));
   }
 
   @Test
@@ -86,14 +90,14 @@ class CentralServerAgencyControllerTest extends BaseApiControllerTest {
     createCentralServer(CS_TEST_CODE2);
 
     stubGet(INNREACH_LOCALSERVERS_URL, xToCodeHeader(CS_TEST_CODE1),
-      "agency-codes/d2ir-local-servers-response-01.json");
+        "agency-codes/d2ir-local-servers-response-01.json");
     stubGet(INNREACH_LOCALSERVERS_URL, xToCodeHeader(CS_TEST_CODE2),
-      "agency-codes/d2ir-local-servers-broken-response.json");
+        "agency-codes/d2ir-local-servers-broken-response.json");
 
     mockMvc.perform(get(AGENCIES_URL))
-      .andExpect(status().isOk())
-      .andExpect(content()
-        .json(readTemplate("cs-agencies-single-server-response.json", cs1.getId(), CS_TEST_CODE1)));
+        .andExpect(status().isOk())
+        .andExpect(content()
+            .json(readTemplate("cs-agencies-single-server-response.json", cs1.getId(), CS_TEST_CODE1)));
   }
 
   @Test
@@ -101,12 +105,12 @@ class CentralServerAgencyControllerTest extends BaseApiControllerTest {
     createCentralServer(CS_TEST_CODE1);
 
     stubFor(WireMock.get(urlEqualTo(INNREACH_LOCALSERVERS_URL))
-      .willReturn(aResponse()
-        .withStatus(HttpStatus.UNAUTHORIZED.value())));
+        .willReturn(aResponse()
+            .withStatus(HttpStatus.UNAUTHORIZED.value())));
 
     mockMvc.perform(get(AGENCIES_URL))
-      .andExpect(status().isOk())
-      .andExpect(content().json(readTemplate("cs-agencies-empty-response.json")));
+        .andExpect(status().isOk())
+        .andExpect(content().json(readTemplate("cs-agencies-empty-response.json")));
   }
 
   @Test
@@ -116,8 +120,8 @@ class CentralServerAgencyControllerTest extends BaseApiControllerTest {
     stubGet(INNREACH_LOCALSERVERS_URL, "agency-codes/d2ir-local-servers-error-response.json");
 
     mockMvc.perform(get(AGENCIES_URL))
-      .andExpect(status().isOk())
-      .andExpect(content().json(readTemplate("cs-agencies-empty-response.json")));
+        .andExpect(status().isOk())
+        .andExpect(content().json(readTemplate("cs-agencies-empty-response.json")));
   }
 
   private CentralServer createCentralServer(String csCode) {
@@ -131,6 +135,21 @@ class CentralServerAgencyControllerTest extends BaseApiControllerTest {
 
   private static Map<String, String> xToCodeHeader(String code) {
     return Map.of(X_TO_CODE, code);
+  }
+
+  private static void stubGet(String url, String responsePath) {
+    stubGet(url, Collections.emptyMap(), responsePath);
+  }
+
+  private static void stubGet(String url, Map<String, String> requestHeaders, String responsePath) {
+    MappingBuilder getBuilder = WireMock.get(urlEqualTo(url));
+
+    requestHeaders.forEach((name, value) -> getBuilder.withHeader(name, equalTo(value)));
+
+    stubFor(getBuilder
+      .willReturn(aResponse()
+          .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
+          .withBodyFile(responsePath)));
   }
 
   private static String readTemplate(String templateFile, Object... params) {
