+++ src/test/java/org/folio/innreach/controller/base/BaseApiControllerTest.java
@@ -1,18 +1,10 @@
 package org.folio.innreach.controller.base;
 
-import static com.github.tomakehurst.wiremock.client.WireMock.aResponse;
-import static com.github.tomakehurst.wiremock.client.WireMock.equalTo;
-import static com.github.tomakehurst.wiremock.client.WireMock.stubFor;
-import static com.github.tomakehurst.wiremock.client.WireMock.urlEqualTo;
 import static com.github.tomakehurst.wiremock.core.WireMockConfiguration.wireMockConfig;
-
-import java.util.Collections;
-import java.util.Map;
 
 import javax.validation.Valid;
 
 import com.github.tomakehurst.wiremock.WireMockServer;
-import com.github.tomakehurst.wiremock.client.MappingBuilder;
 import com.github.tomakehurst.wiremock.client.WireMock;
 import com.github.tomakehurst.wiremock.common.Slf4jNotifier;
 import lombok.extern.log4j.Log4j2;
@@ -24,8 +16,6 @@ import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.Profile;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.context.DynamicPropertyRegistry;
@@ -34,34 +24,33 @@ import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.web.bind.annotation.RestController;
 
 import org.folio.innreach.ModInnReachApplication;
-import org.folio.spring.integration.XOkapiHeaders;
 import org.folio.spring.liquibase.FolioLiquibaseConfiguration;
 import org.folio.tenant.domain.dto.TenantAttributes;
 import org.folio.tenant.rest.resource.TenantApi;
 
 @Log4j2
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,
-  classes = {ModInnReachApplication.class, BaseApiControllerTest.TestTenantController.class})
+    classes = { ModInnReachApplication.class,	BaseApiControllerTest.TestTenantController.class })
 @AutoConfigureMockMvc
-@ActiveProfiles({"test", "testcontainers-pg"})
+@ActiveProfiles({ "test", "testcontainers-pg" })
 public class BaseApiControllerTest {
 
-  @EnableAutoConfiguration(exclude = {FolioLiquibaseConfiguration.class})
+  @EnableAutoConfiguration(exclude = { FolioLiquibaseConfiguration.class })
   @RestController("folioTenantController")
   @Profile("test")
   static class TestTenantController implements TenantApi {
 
     @Override
     public ResponseEntity<String> postTenant(@Valid TenantAttributes tenantAttributes) {
-      return ResponseEntity.ok("OK");
+        return ResponseEntity.ok("OK");
     }
   }
 
   protected static WireMockServer wm =
-    new WireMockServer(wireMockConfig()
-      .dynamicPort()
-      .usingFilesUnderClasspath("wm")
-      .notifier(new Slf4jNotifier(true)));
+      new WireMockServer(wireMockConfig()
+          .dynamicPort()
+          .usingFilesUnderClasspath("wm")
+          .notifier(new Slf4jNotifier(true)));
 
   @Autowired
   protected MockMvc mockMvc;
@@ -89,46 +78,6 @@ public class BaseApiControllerTest {
   @AfterEach
   void tearDown() {
     wm.resetAll();
-  }
-
-  protected static void stubGet(String url, String responsePath) {
-    stubGet(url, Collections.emptyMap(), responsePath);
-  }
-
-  protected static void stubGet(String urlTemplate, String responsePath, Object... pathVariables) {
-    stubGet(String.format(urlTemplate, pathVariables), Collections.emptyMap(), responsePath);
-  }
-
-  protected static void stubGet(String url, Map<String, String> requestHeaders, String responsePath) {
-    MappingBuilder getBuilder = WireMock.get(urlEqualTo(url));
-
-    requestHeaders.forEach((name, value) -> getBuilder.withHeader(name, equalTo(value)));
-
-    stubFor(getBuilder
-      .willReturn(aResponse()
-        .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-        .withHeader(XOkapiHeaders.URL, wm.baseUrl())
-        .withBodyFile(responsePath)));
-  }
-
-  protected static void stubPost(String urlTemplate, String responsePath, Object... pathVariables) {
-    stubPost(String.format(urlTemplate, pathVariables), responsePath);
-  }
-
-  protected static void stubPost(String url, String responsePath) {
-    MappingBuilder postBuilder = WireMock.post(urlEqualTo(url));
-
-    stubFor(postBuilder
-      .willReturn(aResponse()
-        .withHeader(HttpHeaders.CONTENT_TYPE, MediaType.APPLICATION_JSON_VALUE)
-        .withHeader(XOkapiHeaders.URL, wm.baseUrl())
-        .withBodyFile(responsePath)));
-  }
-
-  public HttpHeaders getOkapiHeaders() {
-    HttpHeaders headers = new HttpHeaders();
-    headers.add(XOkapiHeaders.URL, wm.baseUrl());
-    return headers;
   }
 
 }
